pyproject.toml
[tool.poetry]
name = "maps-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
sqlalchemy = "^2.0.36"
geoalchemy2 = "^0.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
isort = "^5.13.2"
pre-commit = "^4.0.1"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
